Organise everything into a package
