+++ src/main/java/Gumball/AddCommand.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 import java.io.IOException;
 
 public class AddCommand extends Command{
