+++ src/main/java/Gumball/Command.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 import java.io.IOException;
 
 public abstract class Command {
