+++ src/main/java/Gumball/DeleteCommand.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 import java.io.IOException;
 
 public class DeleteCommand extends Command{
