+++ src/main/java/Gumball/FileManager.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileWriter;
