+++ src/main/java/Gumball/InputErrorException.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 public class InputErrorException extends Exception {
     public InputErrorException(String str) {
         super(str);
