+++ src/main/java/Gumball/ListCommand.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 import java.io.IOException;
 
 public class ListCommand extends Command{
