+++ src/main/java/Gumball/MarkCommand.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 import java.io.IOException;
 
 public class MarkCommand extends Command{
