+++ src/main/java/Gumball/Parser.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 import java.io.IOException;
 
 public class Parser {
