+++ src/main/java/Gumball/TaskException.java
@@ -1,3 +1,5 @@
+package Gumball;
+
 public class TaskException extends InputErrorException{
     TaskException(String msg) {
         super(msg);
