+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
-Hello! I'm Gumball
+Hello! I'm Gumball.Gumball
 What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
